// 20211SVAC/G23/backend/model/xpath/Enum.ts
export enum Tipos {
    //Nodename unario
    NODENAME = "NODENAME",
    STRING = "STRING",
    NUMBER = "NUMBER",
    ASTERISCO = "ASTERISCO",
    BOOLEANO = "BOOLEANO",

    // Selección
    SELECT_FROM_ROOT = "SELECT_FROM_ROOT",
    SELECT_FROM_CURRENT = "SELECT_FROM_CURRENT",
    SELECT_CURRENT = "SELECT_CURRENT",
    SELECT_PARENT = "SELECT_PARENT",
    SELECT_ATTRIBUTES = "SELECT_ATTRIBUTES",
    SELECT_AXIS = "SELECT_AXIS",

    // Aritméticas
    OPERACION_SUMA = "OPERACION_SUMA",
    OPERACION_RESTA = "OPERACION_RESTA",
    OPERACION_MULTIPLICACION = "OPERACION_MULTIPLICACION",
    OPERACION_DIVISION = "OPERACION_DIVISION",
    OPERACION_MODULO = "OPERACION_MODULO",
    OPERACION_NEGACION_UNARIA = "OPERACION_NEGACION_UNARIA",

    // Relacionales
    RELACIONAL_IGUAL = "RELACIONAL_IGUAL",
    RELACIONAL_DIFERENTE = "RELACIONAL_DIFERENTE",
    RELACIONAL_MENOR = "RELACIONAL_MENOR",
    RELACIONAL_MENORIGUAL = "RELACIONAL_MENORIGUAL",
    RELACIONAL_MAYOR = "RELACIONAL_MAYOR",
    RELACIONAL_MAYORIGUAL = "RELACIONAL_MAYORIGUAL",

    // Logicas
    LOGICA_OR = "LOGICA_OR",
    LOGICA_AND = "LOGICA_AND",

    // Funciones reservadas
    FUNCION_LAST = "FUNCION_LAST",
    FUNCION_POSITION = "FUNCION_POSITION",
    FUNCION_TEXT = "FUNCION_TEXT",
    FUNCION_NODE = "FUNCION_NODE",

    // Predicado
    PREDICATE = "PREDICATE",
    EXPRESION = "EXPRESION",

    // Combinacional
    UNION = "SEVERAL_UNION",

    // Expresiones
    ELEMENTOS = "ELEMENTOS",
    ATRIBUTOS = "ATRIBUTOS",
    TEXTOS = "TEXTOS",
    COMBINADO = "COMBINADO",
    EXCLUDE = "EXCLUDE",

    // Axisnames
    AXIS_ANCESTOR = "ANCESTOR",
    AXIS_ANCESTOR_OR_SELF = "ANCESTOR_OR_SELF",
    AXIS_ATTRIBUTE = "AXIS_ATTRIBUTE",
    AXIS_CHILD = "AXIS_CHILD",
    AXIS_DESCENDANT = "AXIS_DESCENDANT",
    AXIS_DESCENDANT_OR_SELF = "AXIS_DESCENDANT_OR_SELF",
    AXIS_FOLLOWING = "AXIS_FOLLOWING",
    AXIS_FOLLOWING_SIBLING = "AXIS_FOLLOWING_SIBLING",
    AXIS_NAMESPACE = "AXIS_NAMESPACE",
    AXIS_PARENT = "AXIS_PARENT",
    AXIS_PRECEDING = "AXIS_PRECEDING",
    AXIS_PRECEDING_SIBLING = "AXIS_PRECEDING_SIBLING",
    AXIS_SELF = "AXIS_SELF",

    // Default
    NONE = "NONE"
}